Update score type to u_int64_t in position struct and cast player pieces to char in winning move checks

+++ includes/connect4.h
@@ -51,7 +51,7 @@ typedef struct s_pos
 {
 	int x;
 	int y;
-	int score;
+	u_int64_t score;
 } t_pos;
 
 typedef struct s_moves
